requirements.txt
click
pyyaml
xdg
